pyproject.toml
[tool.poetry]
name = "olist-library"
version = "0.1.0"
description = ""
authors = ["Igor Xavier <igoorlimaa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
djangorestframework = "^3.11.0"
django-filter = "^2.3.0"
psycopg2 = "^2.8.5"
django-environ = "^0.4.5"
drf-yasg = "^1.17.1"
django_heroku = "^0.3.1"
gunicorn = "^20.0.4"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
